make `./x.py bench` again
