+++ src/librustc_codegen_ssa/back/rpath/tests.rs
@@ -52,7 +52,7 @@ fn test_minimize2() {
 fn test_rpath_relative() {
     if cfg!(target_os = "macos") {
         let config = &mut RPathConfig {
-            used_crates: Vec::new(),
+            used_crates: &[],
             has_rpath: true,
             is_like_osx: true,
             linker_is_gnu: false,
@@ -64,7 +64,7 @@ fn test_rpath_relative() {
         assert_eq!(res, "@loader_path/../lib");
     } else {
         let config = &mut RPathConfig {
-            used_crates: Vec::new(),
+            used_crates: &[],
             out_filename: PathBuf::from("bin/rustc"),
             get_install_prefix_lib_path: &mut || panic!(),
             has_rpath: true,
